data_columns argument gets modified when calling store.append
I just got bit by passing a class attr (a list) to the store.append function. The list was extended in-place by line 3326 without first making a copy, causing unwanted side effects.

I propose adding a line that would first make a copy of the list before extending it in order to avoid such problems in the future.